change directory and add documentation
