+++ app/decorator.py
@@ -3,6 +3,8 @@ import functools
 # Creamos un decorador que nos ayude a medir el tiempo de ejecución
 # Sin tener que escribirlo para cada función
 # código basado en https://realpython.com/primer-on-python-decorators/#timing-functions
+
+
 def log(func):
     @functools.wraps(func)
     def wrapper(*args, **kwargs):
